# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateTagRequest:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
      slug:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
...
  dto.SubscriptionCountResponse:
    properties:
      authorId:
        type: string
      subscriberCount:
        type: integer
    type: object
  dto.SubscriptionResponse:
    properties:
      author:
        $ref: '#/definitions/dto.UserBriefResponse'
      authorId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      subscriber:
        $ref: '#/definitions/dto.UserBriefResponse'
      subscriberId:
        type: string
    type: object
  dto.TagResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.UpdateTagRequest:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
      slug:
        maxLength: 50
        minLength: 1
        type: string
    type: object
  dto.UserBriefResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Clean Architecture Go API with GORM, Gin, PostgreSQL, Redis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
  /api/v1/authors/{authorId}/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe to receive access to subscriber-only content
      parameters:
      - description: Author ID
        in: path
        name: authorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Subscribe to an author
      tags:
      - Subscriptions
  /api/v1/authors/{authorId}/subscribers:
    get:
      consumes:
      - application/json
      description: Get list of subscribers for an author
      parameters:
      - description: Author ID
        in: path
        name: authorId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get author's subscribers
      tags:
      - Subscriptions
  /api/v1/authors/{authorId}/subscribers/count:
    get:
      consumes:
      - application/json
      description: Get the number of subscribers for an author
      parameters:
      - description: Author ID
        in: path
        name: authorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionCountResponse'
      summary: Get subscriber count
      tags:
      - Subscriptions
  /api/v1/authors/{authorId}/unsubscribe:
    post:
      consumes:
      - application/json
      description: Remove subscription from an author
      parameters:
      - description: Author ID
        in: path
        name: authorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Unsubscribe from an author
      tags:
      - Subscriptions
...
  /api/v1/subscriptions:
    get:
      consumes:
      - application/json
      description: Get list of authors the current user is subscribed to
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get my subscriptions
      tags:
      - Subscriptions
  /api/v1/tags:
    get:
      consumes:
      - application/json
      description: Get all tags
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: List tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Create a new blog tag
      parameters:
      - description: Tag data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a tag
      tags:
      - Tags
  /api/v1/tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a tag
      tags:
      - Tags
    get:
      consumes:
      - application/json
      description: Get a tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get tag by ID
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: Update an existing tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a tag
      tags:
      - Tags
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"